# docs/api/transaction-service.openapi.json
{
  "components" : {
    "schemas" : {
...
      "DisputeNoteRequest" : {
        "properties" : {
          "note" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "note" ],
        "type" : "object"
      },
...
      "Note" : {
        "properties" : {
          "author" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "noteId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Problem" : {
        "description" : "RFC 9457 Problem Details. error, message, path and timestamp repeat title, detail and instance for clients written before the RFC format.",
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "transactionId" : {
            "type" : "string"
          },
          "type" : {
            "example" : "urn:financial:problem:insufficient-funds",
            "format" : "uri",
            "type" : "string"
          },
          "validationErrors" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Field name to message, for validation failures",
            "type" : "object"
          }
        },
        "required" : [ "type", "title", "status" ],
        "type" : "object"
      },
...
      "ReconciliationExceptionAssignmentRequest" : {
        "properties" : {
          "assignedTo" : {
            "type" : "string"
          },
          "expectedVersion" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ReconciliationExceptionNoteRequest" : {
        "properties" : {
          "note" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ReconciliationExceptionNoteResponse" : {
        "properties" : {
          "author" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "noteId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ReconciliationExceptionResponse" : {
        "properties" : {
          "actualAmount" : {
            "type" : "number"
          },
          "assignedTo" : {
            "type" : "string"
          },
          "checkCode" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "deltaAmount" : {
            "type" : "number"
          },
          "detectedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "exceptionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "expectedAmount" : {
            "type" : "number"
          },
          "externalAccountId" : {
            "type" : "string"
          },
          "fingerprint" : {
            "type" : "string"
          },
          "journalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ledgerAccountId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "notes" : {
            "items" : {
              "$ref" : "#/components/schemas/ReconciliationExceptionNoteResponse"
            },
            "type" : "array"
          },
          "resolutionNote" : {
            "type" : "string"
          },
          "runId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ReconciliationExceptionStatusRequest" : {
        "properties" : {
          "expectedVersion" : {
            "format" : "int64",
            "type" : "integer"
          },
          "note" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ReconciliationRunRequest" : {
        "properties" : {
          "businessDate" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "businessDate" ],
        "type" : "object"
      },
      "ReconciliationRunResponse" : {
        "properties" : {
          "businessDate" : {
            "format" : "date",
            "type" : "string"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "criticalExceptions" : {
            "format" : "int32",
            "type" : "integer"
          },
          "reconciliationType" : {
            "type" : "string"
          },
          "requestedBy" : {
            "type" : "string"
          },
          "runId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "totalExceptions" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RiskAlertResponse" : {
        "properties" : {
          "alertId" : {
            "type" : "string"
          },
          "alertType" : {
            "enum" : [ "HIGH_VALUE_TRANSFER", "REPEATED_FAILURES", "RAPID_TRANSFERS", "REVERSAL_HEAVY_ACTIVITY" ],
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "dedupeKey" : {
            "type" : "string"
          },
          "fromAccountId" : {
            "type" : "string"
          },
          "metadata" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "recommendation" : {
            "type" : "string"
          },
          "resolutionNote" : {
            "type" : "string"
          },
          "reviewedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reviewedBy" : {
            "type" : "string"
          },
          "severity" : {
            "enum" : [ "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "OPEN", "REVIEWED", "DISMISSED", "ESCALATED" ],
            "type" : "string"
          },
          "toAccountId" : {
            "type" : "string"
          },
          "transactionId" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "RiskAlertStatusUpdateRequest" : {
        "properties" : {
          "resolutionNote" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "OPEN", "REVIEWED", "DISMISSED", "ESCALATED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RiskCaseNoteRequest" : {
        "properties" : {
          "note" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "RiskCaseResponse" : {
        "properties" : {
          "assignedTo" : {
            "type" : "string"
          },
          "caseId" : {
            "type" : "string"
          },
          "caseNumber" : {
            "type" : "string"
          },
          "claimedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "closedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "linkedAlerts" : {
            "items" : {
              "$ref" : "#/components/schemas/RiskAlertResponse"
            },
            "type" : "array"
          },
          "notes" : {
            "items" : {
              "$ref" : "#/components/schemas/Note"
            },
            "type" : "array"
          },
          "primaryAlertId" : {
            "type" : "string"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH", "CRITICAL" ],
            "type" : "string"
          },
          "resolutionNote" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "OPEN", "IN_REVIEW", "RESOLVED", "CLOSED" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "transactionId" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "RiskCaseStatusUpdateRequest" : {
        "properties" : {
          "resolutionNote" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "OPEN", "IN_REVIEW", "RESOLVED", "CLOSED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TransactionDisputeResponse" : {
        "properties" : {
          "assignedTo" : {
            "type" : "string"
          },
          "claimedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "closedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "disputeId" : {
            "type" : "string"
          },
          "disputeNumber" : {
            "type" : "string"
          },
          "notes" : {
            "items" : {
              "$ref" : "#/components/schemas/Note"
            },
            "type" : "array"
          },
          "reasonCode" : {
            "enum" : [ "UNAUTHORIZED", "DUPLICATE", "INCORRECT_AMOUNT", "SERVICE_NOT_RECEIVED", "OTHER" ],
            "type" : "string"
          },
          "reimbursedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reimbursementAmount" : {
            "type" : "number"
          },
          "reimbursementCurrency" : {
            "type" : "string"
          },
          "reimbursementTransactionId" : {
            "type" : "string"
          },
          "resolutionNote" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "OPEN", "IN_REVIEW", "APPROVED", "DENIED", "CLOSED" ],
            "type" : "string"
          },
          "transactionId" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Transfers, deposits/withdrawals, scheduled transfers, ledger, disputes, risk and Outcome Protection. Errors use RFC 9457 Problem Details.",
    "title" : "Transaction Service API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/admin/reconciliation/exceptions" : {
      "get" : {
        "operationId" : "listExceptions",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "severity",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ReconciliationExceptionResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      }
    },
    "/api/admin/reconciliation/exceptions/{exceptionId}" : {
      "get" : {
        "operationId" : "getException",
        "parameters" : [ {
          "in" : "path",
          "name" : "exceptionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationExceptionResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      }
    },
    "/api/admin/reconciliation/exceptions/{exceptionId}/assignment" : {
      "patch" : {
        "operationId" : "assign",
        "parameters" : [ {
          "in" : "path",
          "name" : "exceptionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReconciliationExceptionAssignmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationExceptionResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      }
    },
    "/api/admin/reconciliation/exceptions/{exceptionId}/notes" : {
      "post" : {
        "operationId" : "addNote_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "exceptionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReconciliationExceptionNoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationExceptionResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      }
    },
    "/api/admin/reconciliation/exceptions/{exceptionId}/status" : {
      "patch" : {
        "operationId" : "updateStatus_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "exceptionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReconciliationExceptionStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationExceptionResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      }
    },
    "/api/admin/reconciliation/runs" : {
      "get" : {
        "operationId" : "listRuns",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ReconciliationRunResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      },
      "post" : {
        "operationId" : "run",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReconciliationRunRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationRunResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "admin-reconciliation-controller" ]
      }
    },
...
    "/api/disputes/admin/{disputeId}/notes" : {
      "post" : {
        "operationId" : "addNote_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "disputeId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DisputeNoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionDisputeResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "transaction-dispute-controller" ]
      }
    },
...
    "/api/risk/alerts/{alertId}/status" : {
      "patch" : {
        "operationId" : "updateStatus_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "alertId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RiskAlertStatusUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RiskAlertResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "risk-alert-controller" ]
      }
    },
...
    "/api/risk/cases/{caseId}/notes" : {
      "post" : {
        "operationId" : "addNote",
        "parameters" : [ {
          "in" : "path",
          "name" : "caseId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RiskCaseNoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RiskCaseResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "risk-case-controller" ]
      }
    },
    "/api/risk/cases/{caseId}/status" : {
      "patch" : {
        "operationId" : "updateStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "caseId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RiskCaseStatusUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RiskCaseResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "risk-case-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Through the gateway",
    "url" : "/transaction-api"
  } ]
}